Make the dataset_build.py oone-single-process
